+++ src/Sentry/Extensibility/DisabledHub.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Threading.Tasks;
-using Sentry.Protocol;
 
 namespace Sentry.Extensibility
 {