+++ test/Sentry.Tests/Internals/DefaultSentryHttpClientFactoryTests.cs
@@ -1,7 +1,5 @@
-using System;
 using System.IO.Compression;
 using System.Linq;
-using System.Net;
 using System.Net.Http;
 using NSubstitute;
 using Sentry.Internal.Http;