+++ test/Sentry.Tests/Protocol/Context/RuntimeTests.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using Sentry.Tests.Protocol;
 using Xunit;
 
 // ReSharper disable once CheckNamespace