+++ test/Sentry.Tests/Protocol/SentryIdTests.cs
@@ -1,5 +1,4 @@
 using System;
-using Sentry.Protocol;
 using Xunit;
 
 namespace Sentry.Tests.Protocol